+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## HEAD
+
+* [vips] Allow forcing a specific loader or saver (@janko)
+
 ## 1.8.0 (2019-02-25)
 
 * [vips] Perform resize-on-load when possible, significantly improving resizing speed (@janko)